Add Responder quiz.

+++ src/main/java/com/QuizMaster/QuizMaster/Repository/QuizzesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuizzesRepository extends JpaRepository<Quizzes, Integer> {
+public interface QuizzesRepository extends JpaRepository<Quizzes, Long> {
     // Aqui pode criar métodos personalizados depois, tipo buscar por categoria ou nível
 }
